+++ src/compiler/command_line_parser/mod.rs
@@ -1,0 +1,4 @@
+mod lines_0_1000;
+pub use lines_0_1000::*;
+mod lines_1000_end;
+pub use lines_1000_end::*;
